Feature auth

+++ backend/users/views.py
@@ -3,6 +3,7 @@ from sqlalchemy import orm
 import sqlalchemy
 import api
 import flask
+import auth
 
 users_router = flask.Blueprint('users_urls', 'users')
 
@@ -21,12 +22,12 @@ def get_user(user_id: int):
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
     if user is None:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
     return flask.jsonify({
         'error': False,
@@ -55,7 +56,7 @@ def get_users():
     if (page_size < 1 or page_size > 100):
         return flask.jsonify({
             'error': True,
-            'detail': f'Invalid page size: { page_size }'
+            'detail': f'Invalid page size: {page_size}'
         }), 400
 
     page = flask.request.args.get('page', 1, type=int)
@@ -63,7 +64,7 @@ def get_users():
     if (page < 1):
         return flask.jsonify({
             'error': True,
-            'detail': f'Invalid page: { page }',
+            'detail': f'Invalid page: {page}',
         }), 400
 
     try:
@@ -144,7 +145,7 @@ def change_password(user_id: int):
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
 
     if user is None:
@@ -183,6 +184,12 @@ def edit_user_settings(user_id: int):
         nickname (:obj:`str`, optional): New nickname of the new user.
     '''
 
+    if auth.verify_token(auth.auth.current_user()).id != user_id:
+        return flask.jsonify({
+            'error': True,
+            'detail': 'Can\'t edit data of another user'
+        }), 401
+
     email = flask.request.args.get('email', None, type=str)
     nickname = flask.request.args.get('nickname', None, type=str)
 
@@ -191,7 +198,7 @@ def edit_user_settings(user_id: int):
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
 
     if user is None:
@@ -238,6 +245,13 @@ def edit_user_settings(user_id: int):
 @users_router.route('/<int:user_id>', methods=['DELETE'])
 def delete_user(user_id: int):
     '''Deletes user by given id'''
+
+    if auth.verify_token(auth.auth.current_user()).id != user_id:
+        return flask.jsonify({
+            'error': True,
+            'detail': 'Can\'t edit data of another user'
+        }), 401
+
     try:
         models.User.delete_user(user_id)
         return flask.jsonify({
@@ -251,14 +265,15 @@ def delete_user(user_id: int):
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
 
 @users_router.route('/<int:user_id>', methods=['GET'])
 def get_time_settings(user_id: int):
     '''Gets time settings by an id.
 
-    Returns 3 time settings by given id (int). If user not found returns 404 error.
+    Returns 3 time settings by given id (int).
+    If user not found returns 404 error.
 
     Args:
         user_id(int): user\'s id
@@ -269,12 +284,12 @@ def get_time_settings(user_id: int):
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
     if user is None:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
     return flask.jsonify({
         'error': False,
@@ -292,30 +307,38 @@ def edit_user_sttings(user_id: int):
     Edits user by given id.
 
     Args:
-        time_for_reading (:obj:`int`, optional): New time for reading setting of the user.
-        time_for_solving (:obj:`int`, optional): New time for solving setting of the user.
-        time_for_typing (:obj:`int`, optional): New time for typing setting of the user.
+        time_for_reading (:obj:`int`, optional):
+        New time for reading setting of the user.
+        time_for_solving (:obj:`int`, optional):
+        New time for solving setting of the user.
+        time_for_typing (:obj:`int`, optional):
+        New time for typing setting of the user.
     '''
 
-    time_for_reading = flask.request.args.get('time_for_reading', None, type=int)
-    time_for_solving = flask.request.args.get('time_for_solving', None, type=int)
-    time_for_typing = flask.request.args.get('time_for_typing', None, type=int)
+    time_for_reading = flask.request.args.get('time_for_reading',
+                                              None, type=int)
+    time_for_solving = flask.request.args.get('time_for_solving',
+                                              None, type=int)
+    time_for_typing = flask.request.args.get('time_for_typing',
+                                             None, type=int)
 
     try:
         user = models.User.get_user(user_id)
     except OverflowError:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }',
+            'detail': f'Could not find user with id {user_id}',
         }), 404
 
     if user is None:
         return flask.jsonify({
             'error': True,
-            'detail': f'Could not find user with id { user_id }'
+            'detail': f'Could not find user with id {user_id}'
         }), 404
 
-    user.update_user_settings(time_for_reading, time_for_solving, time_for_typing)
+    user.update_user_settings(time_for_reading,
+                              time_for_solving,
+                              time_for_typing)
     return flask.jsonify({
         'error': False,
         'time_for_reading': user.time_for_reading,
